requirements.txt
certifi==2020.12.5
chardet==4.0.0
decorator==4.4.2
docopt==0.6.2
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.3
moviepy==1.0.3
numpy==1.20.2
Pillow==8.1.2
proglog==0.1.9
requests==2.25.1
tqdm==4.59.0
urllib3==1.26.4
webvtt-py==0.4.6
yattag==1.14.0
youtube-dl==2021.4.1
